requirements.txt
selenium>=4.30.0
webdriver-manager>=4.0.0
pytest>=8.0.0
allure-pytest>=2.13.0
